Water Bottle Class Completed, Four Tests Passing

+++ src/main/java/Calculator.java
@@ -1,21 +1,21 @@
 public class Calculator {
 
 
-    public int multiply(int num1, int num2){
+    public double multiply(double num1, double num2){
         return( num1 * num2);
     }
 
 
-    public int divide(int num1, int num2){
+    public double divide(double num1, double num2){
         return (num1 / num2);
     }
 
 
-    public int add(int num1, int num2){
+    public double add(double num1, double num2){
         return (num1 + num2);
     }
 
-    public int subtract(int num1, int num2){
+    public double subtract(double num1, double num2){
         return (num1 - num2);
     }
 }
